Remove "Sort Crimes" functionality
When the Crimes were being primarily stored in an ArrayList, sorting the
Crimes in the RecyclerView only required sorting that ArrayList before
updating the Adapter. Now, the Crimes need to be sorted in the database
to realistically be sorted while displayed. A sortCrimes() method was
attempted in CrimeAdapter, but any outer call to updateUi() displayed
the Crimes in their actual order again.